Issue #2930486 by Kingdutch: Update patch for group relation deletion bug
The patch that was previously applied for fixing group deletion
didn't fix the issue but only muted it in a specific instance.
This commit updates the patch to the one that was committed and has
properly addressed the bug.

This is needed because when deleting a user and removing the issue
post content, its group relation becomes the a blocking issue.